requirements.txt
azure-devops==6.0.0b4
certifi==2021.10.8
charset-normalizer==2.0.9
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
smmap==5.0.0
typing-extensions==4.0.1
urllib3==1.26.7
